Update rate in restaurant

+++ RaterestaurantMVC.Application/ViewModels/Opinion/NewOpinionVm.cs
@@ -12,7 +12,7 @@ namespace RaterestaurantMVC.Application.ViewModels.Opinion
     {
         public int Id { get; set; }
         public string Content { get; set; }
-        public double Rate { get; set; }
+        public int Rate { get; set; }
         public int UserId { get; set; }
         public int RestaurantId { get; set; }
